+++ matholymp/scripts/mo_document_generate.py
@@ -36,11 +36,11 @@ when logged in administratively), countries.csv and flags/ and photos/
 directories from unpacking the corresponding .zip files, and generates
 output in an out/ subdirectory, unless options are used to select
 other directories.  The known types of document are 'badge',
-'desk-label', 'award-certificate', 'participation-certificate',
-'paper', 'language-list', 'coord-form' and 'scores-commands'.  The
-identifiers passed along with document types vary; any identifier may
-be used for document types that do not depend on such an identifier,
-and 'all' is suitable for most cases.
+'visa-letter', 'desk-label', 'award-certificate',
+'participation-certificate', 'paper', 'language-list', 'coord-form'
+and 'scores-commands'.  The identifiers passed along with document
+types vary; any identifier may be used for document types that do not
+depend on such an identifier, and 'all' is suitable for most cases.
 """
 
 import argparse
@@ -105,6 +105,8 @@ def main():
     bg = cmdline_data['background']
     if cmdline_data['type'] == 'badge':
         docgen.generate_badges(gen_id, bg)
+    elif cmdline_data['type'] == 'visa-letter':
+        docgen.generate_visa_letter(gen_id)
     elif cmdline_data['type'] == 'desk-label':
         docgen.generate_desk_labels(gen_id)
     elif cmdline_data['type'] == 'award-certificate':